use int for expiry, is easier to handle and sufficient for the purpose
git-svn-id: https://svn.exedio.com/cope/trunk/persistence@11392 e7d4fc99-c606-0410-b9bf-843393a9eab7

+++ runtime/patternsrc/com/exedio/cope/pattern/PasswordRecovery.java
@@ -35,7 +35,7 @@ public final class PasswordRecovery extends Pattern
 	private static final long NOT_A_TOKEN = 0l;
 	
 	private final Hash password;
-	private final long expiryMillis;
+	private final int expiryMillis;
 	private final LongField token = new LongField().defaultTo(NOT_A_TOKEN);
 	private final DateField date = new DateField().optional();
 	private final SecureRandom random = new SecureRandom();
@@ -45,7 +45,7 @@ public final class PasswordRecovery extends Pattern
 		this(password, 15*60*1000); // 15 minutes
 	}
 	
-	public PasswordRecovery(final Hash password, final long expiryMillis)
+	public PasswordRecovery(final Hash password, final int expiryMillis)
 	{
 		this.password = password;
 		this.expiryMillis = expiryMillis;
@@ -63,12 +63,12 @@ public final class PasswordRecovery extends Pattern
 		return password;
 	}
 	
-	public long getExpiryMillis()
+	public int getExpiryMillis()
 	{
 		return expiryMillis;
 	}
 	
-	public long getExpiryMinutes()
+	public int getExpiryMinutes()
 	{
 		return expiryMillis / (60*1000);
 	}
